Did alotta changes...

+++ src/main/java/com/unionsg/xaccounting/entity/ChartOfAccountClearTo_ENTITY.java
@@ -34,7 +34,7 @@ public class ChartOfAccountClearTo_ENTITY {
     // there should be a function to be creating the codes
     // maybe by incrementing or by generating random codes.
     @Column(name = "clear_to_code")
-    private int clearToCode;
+    private Long clearToCode;
 
     @OneToMany(mappedBy = "coaClearTo", cascade = CascadeType.ALL)
     private List<AccountEntity> accounts;
